allmusicgrabber/api/requirements.txt
Werkzeug==2.2.2
flask
bs4
requests
flask-oauthlib
python-dotenv

